Bugfix account.unit to refresh account page when you change it
this also start splitting a bit the AccountPage component

as CalculateBalance will memoize things, we want inner part to be potentially re-render if they depend on other things, we do this by connecting them to redux. redux idea is to connect leafs instead of connecting the whole tree when possible, so this fit this paradigm to me.
